Renamed "components" file to be private ("_components")
