+++ apps/scoop-portal/src/app/_components/Footer.js
